voices/requirements.txt
aider-chat==0.12.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ConfigArgParse==1.7
diskcache==5.6.1
exceptiongroup==1.1.3
ffmpeg==1.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
iniconfig==2.0.0
jsonschema==4.17.3
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
networkx==3.1
numpy==1.24.3
openai==0.27.6
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.38
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.15.1
pyrsistent==0.19.3
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.30.0
rich==13.3.5
scipy==1.10.1
smmap==5.0.0
sounddevice==0.4.6
soundfile==0.12.1
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.2
wcwidth==0.2.6
yarl==1.9.2
